Backport 33d9a857308eed53e06b448691910bc8aa2f8fc9
jdk11 also has the bug of JDK-8307572, so I propose to backport the fix
to jdk11u. This is not a clean backport. Main changes:

1) Definition of reg_class v4_reg..v7_reg, and operand
vRegD_V4..vRegD_V7 are missing in jdk11u aarch64.ad, which was
introduced in JDK-8214527. Added those definitions in this patch, as we
need to claim them to be killed in match rules.

2) JDK-8274243 (not a valid backport for 11u) changed
MacroAssembler::encode_iso_array() a lot, but the bug of invalid use of
v4/v5 still exists before and after that patch. Fixed the conflicts by
using 11u naming convention.

No new jtreg test failure with both release and fastdebug build. Test
case added in JDK-8307572 passes now, which fails without the backport.